Fixed all of Jason's comments: - Migrated all profile specific code and functionality to api_contributors - /api/has-workgroup -> /api/workgroup/exists - Refactored tests - Fixed error response logic
