+++ components/centraldashboard/public/components/main-page.pug
@@ -1,4 +1,4 @@
-iron-ajax(auto, url='/api/has-workgroup', handle-as='json',
+iron-ajax(auto, url='/api/workgroup/exists', handle-as='json',
     on-response='_onHasWorkgroupResponse', loading='{{pageLoading}}')
 iron-ajax#envInfo(auto='[[_shouldFetchEnv]]', url='/api/env-info', handle-as='json',
     on-response='_onEnvInfoResponse')
